feat: cardPayments basic crud

+++ backend/src/main/java/com/bia/dev_bank/dto/accountDTOs/AccountRequest.java
@@ -2,7 +2,9 @@ package com.bia.dev_bank.dto.accountDTOs;
 
 import com.bia.dev_bank.entity.enums.AccountType;
 
+import java.math.BigDecimal;
+
 public record AccountRequest(
                              AccountType accountType,
-                             double currentBalance) {
+                             BigDecimal currentBalance) {
 }
