+++ backend/src/main/java/com/bia/dev_bank/dto/accountDTOs/AccountResponse.java
@@ -3,11 +3,13 @@ package com.bia.dev_bank.dto.accountDTOs;
 import com.bia.dev_bank.entity.Account;
 import com.bia.dev_bank.entity.enums.AccountType;
 
+import java.math.BigDecimal;
+
 public record AccountResponse(
         String accountNumber,
         String customerName,
         AccountType accountType,
-        double currentBalance
+        BigDecimal currentBalance
 ) {
     public AccountResponse(Account account) {
         this(
